it is exausting socktes, fix later

+++ Unichain.P2P/Packets/RequestBuilder.cs
@@ -4,7 +4,7 @@
 /// A class to craft new requests.
 /// </summary>
 public class RequestBuilder {
-    private int protocolVersion;
+    private byte protocolVersion;
     private RequestMethod method;
     private Route? route;
     private Address? sender;
@@ -23,7 +23,7 @@ public class RequestBuilder {
     /// Defines the protocol version that the request will use
     /// </summary>
     /// <param name="protocolVersion">The protocol version</param>
-    public RequestBuilder WithProtocolVersion(int protocolVersion) {
+    public RequestBuilder WithProtocolVersion(byte protocolVersion) {
         this.protocolVersion = protocolVersion;
         return this;
     }
